+++ src/configs/eslint-comments.ts
@@ -1,8 +1,7 @@
 // @ts-expect-error no typings
 import eslintPluginEslintComments from '@eslint-community/eslint-plugin-eslint-comments';
 import {ERROR} from '../constants';
-import type {ConfigSharedOptions, FlatConfigEntry} from '../eslint';
-import {ConfigEntryBuilder} from '../eslint';
+import {ConfigEntryBuilder, type ConfigSharedOptions, type FlatConfigEntry} from '../eslint';
 import type {InternalConfigOptions} from './index';
 
 export interface EslintCommentsEslintConfigOptions