+++ src/configs/package-json.ts
@@ -1,8 +1,12 @@
 import eslintPluginPackageJsonRecommendedConfig from 'eslint-plugin-package-json/configs/recommended';
 import jsoncEslintParser from 'jsonc-eslint-parser';
 import {ERROR, GLOB_PACKAGE_JSON} from '../constants';
-import type {ConfigSharedOptions, FlatConfigEntry, GetRuleOptions} from '../eslint';
-import {ConfigEntryBuilder} from '../eslint';
+import {
+  ConfigEntryBuilder,
+  type ConfigSharedOptions,
+  type FlatConfigEntry,
+  type GetRuleOptions,
+} from '../eslint';
 import type {InternalConfigOptions} from './index';
 
 const DEFAULT_FILES = [GLOB_PACKAGE_JSON];