+++ src/configs/prefer-arrow-functions.ts
@@ -1,8 +1,7 @@
 // @ts-expect-error no typings
 import eslintPluginPreferArrowFunctions from 'eslint-plugin-prefer-arrow-functions';
 import {WARNING} from '../constants';
-import type {ConfigSharedOptions, FlatConfigEntry} from '../eslint';
-import {ConfigEntryBuilder} from '../eslint';
+import {ConfigEntryBuilder, type ConfigSharedOptions, type FlatConfigEntry} from '../eslint';
 import type {InternalConfigOptions} from './index';
 
 export interface PreferArrowFunctionsEslintConfigOptions