+++ src/configs/promise.ts
@@ -1,8 +1,7 @@
 // @ts-expect-error no typings
 import eslintPluginPromise from 'eslint-plugin-promise';
 import {ERROR, WARNING} from '../constants';
-import type {ConfigSharedOptions, FlatConfigEntry} from '../eslint';
-import {ConfigEntryBuilder} from '../eslint';
+import {ConfigEntryBuilder, type ConfigSharedOptions, type FlatConfigEntry} from '../eslint';
 import type {InternalConfigOptions} from './index';
 
 export interface PromiseEslintConfigOptions extends ConfigSharedOptions<'promise'> {}