+++ src/configs/security.ts
@@ -1,7 +1,6 @@
 import eslintPluginSecurity from 'eslint-plugin-security';
 import {ERROR, OFF, WARNING} from '../constants';
-import type {ConfigSharedOptions, FlatConfigEntry} from '../eslint';
-import {ConfigEntryBuilder} from '../eslint';
+import {ConfigEntryBuilder, type ConfigSharedOptions, type FlatConfigEntry} from '../eslint';
 import type {InternalConfigOptions} from './index';
 
 export interface SecurityEslintConfigOptions extends ConfigSharedOptions<'security'> {}