+++ src/configs/sonar.ts
@@ -1,7 +1,6 @@
 import eslintPluginSonar from 'eslint-plugin-sonarjs';
 import {OFF} from '../constants';
-import type {ConfigSharedOptions, FlatConfigEntry} from '../eslint';
-import {ConfigEntryBuilder} from '../eslint';
+import {ConfigEntryBuilder, type ConfigSharedOptions, type FlatConfigEntry} from '../eslint';
 import type {InternalConfigOptions} from './index';
 
 export interface SonarEslintConfigOptions extends ConfigSharedOptions<'sonarjs'> {}