+++ src/configs/toml.ts
@@ -1,8 +1,12 @@
 import eslintPluginToml from 'eslint-plugin-toml';
 import tomlEslintParser from 'toml-eslint-parser';
 import {ERROR, GLOB_TOML, OFF} from '../constants';
-import type {ConfigSharedOptions, FlatConfigEntry, GetRuleOptions} from '../eslint';
-import {ConfigEntryBuilder} from '../eslint';
+import {
+  ConfigEntryBuilder,
+  type ConfigSharedOptions,
+  type FlatConfigEntry,
+  type GetRuleOptions,
+} from '../eslint';
 import type {InternalConfigOptions} from './index';
 
 const DEFAULT_FILES = [GLOB_TOML];